refactor(module): extract `commonTokenizer`
